tools/requirements.txt
flask>=3.0
click>=8.0
ruamel.yaml>=0.18
